Rendición de comisiones terminado

+++ src/PalcoNet/DAO/ItemDAO.cs
@@ -9,8 +9,8 @@ namespace DAO
     public class ItemDAO
     {
         public Int32 Item_ID { get; set; }
-		public Int32 Item_Monto { get; set; }
-		public Int32 Item_Cantidad { get; set; }
+		public Decimal Item_Monto { get; set; }
+		public Decimal Item_Cantidad { get; set; }
 		public String Item_Descripcion { get; set; }
 		public Int32 Factura_ID { get; set; }
         public Int32 Compra_ID { get; set; }
